move MsgSet to msg_pool.rs
