+++ blockchain/message_pool/src/msgpool/mod.rs
@@ -10,6 +10,7 @@ pub(crate) mod utils;
 use super::errors::Error;
 use crate::msg_chain::{MsgChain, MsgChainNode};
 use crate::msg_pool::add_helper;
+use crate::msg_pool::MsgSet;
 use crate::msgpool::selection::{add, rm};
 use crate::provider::Provider;
 use address::Address;
@@ -24,7 +25,7 @@ use log::{error, warn};
 use lru::LruCache;
 use message::{Message, SignedMessage, UnsignedMessage};
 use networks::BLOCK_DELAY_SECS;
-use num_bigint::{BigInt, Integer};
+use num_bigint::BigInt;
 use std::collections::{HashMap, HashSet};
 use std::{borrow::BorrowMut, cmp::Ordering};
 use tokio::sync::broadcast::{Receiver as Subscriber, Sender as Publisher};
@@ -38,87 +39,6 @@ const BASE_FEE_LOWER_BOUND_FACTOR: i64 = 10;
 const REPUB_MSG_LIMIT: usize = 30;
 const PROPAGATION_DELAY_SECS: u64 = 6;
 const REPUBLISH_INTERVAL: u64 = 10 * BLOCK_DELAY_SECS + PROPAGATION_DELAY_SECS;
-
-/// Simple struct that contains a hashmap of messages where k: a message from address, v: a message
-/// which corresponds to that address.
-#[derive(Clone, Default, Debug)]
-pub struct MsgSet {
-    msgs: HashMap<u64, SignedMessage>,
-    next_sequence: u64,
-    required_funds: BigInt,
-}
-
-impl MsgSet {
-    /// Generate a new MsgSet with an empty hashmap and setting the sequence specifically.
-    pub fn new(sequence: u64) -> Self {
-        MsgSet {
-            msgs: HashMap::new(),
-            next_sequence: sequence,
-            required_funds: Default::default(),
-        }
-    }
-
-    /// Add a signed message to the MsgSet. Increase next_sequence if the message has a
-    /// sequence greater than any existing message sequence.
-    pub fn add(&mut self, m: SignedMessage) -> Result<(), Error> {
-        if self.msgs.is_empty() || m.sequence() >= self.next_sequence {
-            self.next_sequence = m.sequence() + 1;
-        }
-        if let Some(exms) = self.msgs.get(&m.sequence()) {
-            if m.cid()? != exms.cid()? {
-                let premium = exms.message().gas_premium();
-                let rbf_denom = BigInt::from(RBF_DENOM);
-                let min_price = premium + ((premium * RBF_NUM).div_floor(&rbf_denom)) + 1u8;
-                if m.message().gas_premium() <= &min_price {
-                    return Err(Error::GasPriceTooLow);
-                }
-            } else {
-                return Err(Error::DuplicateSequence);
-            }
-        }
-        self.msgs.insert(m.sequence(), m);
-        Ok(())
-    }
-
-    /// Removes message with the given sequence. If applied, update the set's next sequence.
-    pub fn rm(&mut self, sequence: u64, applied: bool) {
-        let m = if let Some(m) = self.msgs.remove(&sequence) {
-            m
-        } else {
-            if applied && sequence >= self.next_sequence {
-                self.next_sequence = sequence + 1;
-                while self.msgs.get(&self.next_sequence).is_some() {
-                    self.next_sequence += 1;
-                }
-            }
-            return;
-        };
-        self.required_funds -= m.required_funds();
-
-        // adjust next sequence
-        if applied {
-            // we removed a (known) message because it was applied in a tipset
-            // we can't possibly have filled a gap in this case
-            if sequence >= self.next_sequence {
-                self.next_sequence = sequence + 1;
-            }
-            return;
-        }
-        // we removed a message because it was pruned
-        // we have to adjust the sequence if it creates a gap or rewinds state
-        if sequence < self.next_sequence {
-            self.next_sequence = sequence;
-        }
-    }
-
-    fn get_required_funds(&self, sequence: u64) -> BigInt {
-        let required_funds = self.required_funds.clone();
-        match self.msgs.get(&sequence) {
-            Some(m) => required_funds - m.required_funds(),
-            None => required_funds,
-        }
-    }
-}
 
 /// Attempt to get a signed message that corresponds to an unsigned message in bls_sig_cache.
 async fn recover_sig(
