+++ blockchain/message_pool/src/msgpool/selection.rs
@@ -10,8 +10,8 @@ use super::provider::Provider;
 use super::{create_message_chains, msg_pool::MessagePool};
 use crate::msg_chain::MsgChain;
 use crate::msg_pool::remove;
+use crate::msg_pool::MsgSet;
 use crate::Error;
-use crate::MsgSet;
 use address::Address;
 use async_std::sync::{Arc, RwLock};
 use blocks::Tipset;
